Introduce a system-private Java interface to PerspectiveService
PerspectiveManager lets Java code interact with PerspectiveService
through JNI calls to an IPerspectiveService Binder proxy.

This is currently implemented in the simplest way possible. Each
instance of PerspectiveManager will maintain it's own native client.
This will only be created once in System UI, which is always running,
so this should be good for now. If this class is in multiple places it
may be worthwhile using a singleton so that a single proxy can be shared
throughout the calling process.

[Error handling]

We return false in all error cases, including when the remote service
does not exist. There is no retry if we cannot connect to the remote.

[Tests]
    - remote service cannot be found through service_manager
    - remote service unexpectedly dies

This is a critical step in building Maru UI so that users can monitor
and control their desktop perspective.

 Conflicts:
	core/jni/Android.mk
	core/jni/AndroidRuntime.cpp

Signed-off-by: NoelMacwan <yulespencer@gmail.com>