Maru UI: add a desktop tile to Quick Settings
This adds a desktop tile to the quick settings swipe down. It is
a simple toggle style setting: tap to start/stop your desktop.

We currently only show the tile if an HDMI screen is connected or
the desktop is currently running in the background.

The underlying mechanism for controlling the desktop state is through
PerspectiveManager.

 Conflicts:
	packages/SystemUI/res/values/config.xml
	packages/SystemUI/res/values/strings.xml
	packages/SystemUI/src/com/android/systemui/statusbar/phone/QSTileHost.java

Signed-off-by: NoelMacwan <yulespencer@gmail.com>