Introduce dynamic screen mirroring for Android
This adds dynamic mirroring support within DisplayManagerService and
related Android internal APIs for system apps.

To understand how this works, first some background on display
management within Android. DisplayManagerService basically maps
logical displays to "real physical displays", which are called display
devices. Each logical display is uniquely recognized by its display ID.
Each display device is mapped to show a logical display by setting the
display device's layerstack (via SurfaceFlinger) to be the logical
display's display ID.

Soo, the way it is right now.
Stock Android mirrors the logical display DEFAULT_DISPLAY
(the device's built-in screen) to a display device iff that device
has not explicitly requested to only show its own content, and it
either doesn't have a logical display assigned or that display is not
showing anything. It's pretty hidden honestly.

This commit introduces a clear global flag that holds whether mirroring
is enabled or not in DisplayManagerService. DisplayManagerService checks this flag
before attempting to mirror.

We provide Binder APIs to enable, disable, and check the status of this state.
When the flag has changed to a new state, we simply trigger a re-configuration
of display devices.

NOTE: As the DEFAULT_EXTERNAL_DISPLAY is not managed by the window manager,
it never makes use of hasOwnContent; it's always false. We are silently using
that to allow the dynamic mirroring even when there is content on the display.

 Conflicts:
	services/core/java/com/android/server/display/DisplayDeviceInfo.java

Signed-off-by: NoelMacwan <yulespencer@gmail.com>